requirements.txt
agate==1.6.3
attrs==22.2.0
Babel==2.11.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
databricks-cli==0.17.4
databricks-connect==11.3.0b0
databricks-sql-connector==2.2.1
dbt-core==1.3.1
dbt-databricks==1.3.2
dbt-extractor==0.4.1
dbt-spark==1.3.0
future==0.18.2
hologram==0.0.15
ibm-db==3.1.4
idna==3.4
isodate==0.6.1
Jinja2==3.1.2
jsonschema==3.2.0
leather==0.3.4
Logbook==1.5.3
lz4==4.0.2
MarkupSafe==2.1.1
mashumaro==3.0.4
minimal-snowplow-tracker==0.0.2
msgpack==1.0.4
networkx==2.8.8
numpy==1.23.4
oauthlib==3.2.2
packaging==21.3
pandas==1.5.2
parsedatetime==2.4
pathspec==0.9.0
py4j==0.10.9.5
pyarrow==10.0.1
pycparser==2.21
PyJWT==2.6.0
pyodbc==4.0.35
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-slugify==7.0.0
pytimeparse==1.1.8
pytz==2022.7
PyYAML==6.0
requests==2.28.1
six==1.16.0
sqlparams==5.0.0
sqlparse==0.4.3
tabulate==0.9.0
text-unidecode==1.3
thrift==0.16.0
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
